+++ packages/client-web/__tests__/unit/web_client.test.ts
@@ -3,20 +3,22 @@ import { createClient } from '../../src'
 
 describe('[Web] createClient', () => {
   it('throws on incorrect "url" config value', () => {
-    expect(() => createClient({ url: 'foo' })).toThrowError(
-      'ClickHouse URL is malformed.'
+    expect(() => createClient({ url: 'foo' })).toThrow(
+      jasmine.objectContaining({
+        message: jasmine.stringContaining('ClickHouse URL is malformed.'),
+      })
     )
   })
 
   it('should not mutate provided configuration', async () => {
     const config: BaseClickHouseClientConfigOptions = {
-      url: 'http://localhost',
+      url: 'https://localhost:8443',
     }
     createClient(config)
     // initial configuration is not overridden by the defaults we assign
     // when we transform the specified config object to the connection params
     expect(config).toEqual({
-      url: 'http://localhost',
+      url: 'https://localhost:8443',
     })
   })
 })